requirements.txt
anyio==3.4.0
certifi==2021.10.8
charset-normalizer==2.0.9
docopt==0.6.2
fastapi==0.70.1
fastapi-login==1.7.3
idna==3.3
numpy==1.21.2
passlib==1.7.4
pipreqs==0.4.11
pydantic==1.8.2
PyJWT==2.3.0
python-multipart==0.0.5
requests==2.26.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.1
urllib3==1.26.7
yarg==0.1.9
